+++ test/test_ops.jl
@@ -180,7 +180,7 @@ end # of testsection
     @test minortranspose(minortranspose(AA_sym)) ≈ AA_sym
 
     @test majortranspose((@inferred majortranspose(AA))::Tensor{4, dim, T}) ≈ AA
-    @test majortranspose((@inferred majortranspose(AA_sym))::Tensor{4, dim, T}) ≈ AA_sym
+    @test majortranspose((@inferred majortranspose(AA_sym))::SymmetricTensor{4, dim, T}) ≈ AA_sym
     @test AA[i,j,k,l] ≈ majortranspose(AA)[k,l,i,j]
     @test AA_sym[i,j,k,l] ≈ majortranspose(AA_sym)[k,l,i,j]